Use same token receival
